+++ src/globals.js
@@ -1,31 +1,27 @@
 import Akili from './akili.js';
 
+const obj = {};
+
 /**
  * Globals to use the custom variables in scope expressions
  * 
  * {@link https://akilijs.com/docs/scope#docs_globals}
  */
-const globals = new Proxy({}, {
+const globals = new Proxy(obj, {
   get: (target, key) => {
-    const tag = `globals.${key}`;
-
-    if(Akili.__evaluation) {
-      Akili.__evaluation.component.__addTag(tag, Akili.__evaluation.node);
-      return target[key];
-    }
+    if(key == '__target') {
+      return obj
+    }   
+    
+    return target[key];
   },
   set: (target, key, value) => {
-    const tag = `globals.${key}`; 
-    const prev = target[key];
-    target[key] = Akili.wrap(value);    
-    prev !== value && Akili.evaluateTag(tag);
+    target[key] = Akili.wrap(value, { tag: key });    
     return true;
   },
-  deleteProperty: (target, key) => {
-    const tag = `globals.${key}`;    
-    Akili.removeTag(tag);
+  deleteProperty: (target, key) => {   
+    Akili.removeTag(key);
     delete target[key];
-    Akili.evaluateTag(tag);
     return true;
   }
 });